Avoid 32-bit conversion errors

+++ ref/PQCgenKAT_sign.c
@@ -34,7 +34,7 @@ main()
     unsigned char       msg[3300];
     unsigned char       entropy_input[48];
     unsigned char       *m, *sm, *m1;
-    unsigned long long  mlen, smlen, mlen1;
+    size_t              mlen, smlen, mlen1;
     int                 count;
     int                 done;
     unsigned char       pk[CRYPTO_PUBLICKEYBYTES], sk[CRYPTO_SECRETKEYBYTES];
@@ -61,7 +61,7 @@ main()
         randombytes(seed, 48);
         fprintBstr(fp_req, "seed = ", seed, 48);
         mlen = (unsigned long long int)(33*(i+1));
-        fprintf(fp_req, "mlen = %llu\n", mlen);
+        fprintf(fp_req, "mlen = %zu\n", mlen);
         randombytes(msg, mlen);
         fprintBstr(fp_req, "msg = ", msg, mlen);
         fprintf(fp_req, "pk =\n");
@@ -97,12 +97,12 @@ main()
         randombytes_init(seed, NULL);
 
         if ( FindMarker(fp_req, "mlen = ") )
-            ret_val = fscanf(fp_req, "%llu", &mlen);
+            ret_val = fscanf(fp_req, "%zu", &mlen);
         else {
             printf("ERROR: unable to read 'mlen' from <%s>\n", fn_req);
             return KAT_DATA_ERROR;
         }
-        fprintf(fp_rsp, "mlen = %llu\n", mlen);
+        fprintf(fp_rsp, "mlen = %zu\n", mlen);
 
         m = (unsigned char *)calloc(mlen, sizeof(unsigned char));
         m1 = (unsigned char *)calloc(mlen+CRYPTO_BYTES, sizeof(unsigned char));
@@ -126,7 +126,7 @@ main()
             printf("crypto_sign returned <%d>\n", ret_val);
             return KAT_CRYPTO_FAILURE;
         }
-        fprintf(fp_rsp, "smlen = %llu\n", smlen);
+        fprintf(fp_rsp, "smlen = %zu\n", smlen);
         fprintBstr(fp_rsp, "sm = ", sm, smlen);
         fprintf(fp_rsp, "\n");
 
@@ -136,7 +136,7 @@ main()
         }
 
         if ( mlen != mlen1 ) {
-            printf("crypto_sign_open returned bad 'mlen': Got <%llu>, expected <%llu>\n", mlen1, mlen);
+            printf("crypto_sign_open returned bad 'mlen': Got <%zu>, expected <%zu>\n", mlen1, mlen);
             return KAT_CRYPTO_FAILURE;
         }
